fix: simplify mocked api code
