+++ src/app/api/jhipster-mocks/_helpers/db.ts
