+++ src/app/api/jhipster-mocks/account/reset-password/finish/route.ts
@@ -1,11 +1,13 @@
+import bcrypt from 'bcrypt';
 import { NextResponse } from 'next/server';
 import { z } from 'zod';
 
-import { resetPasswordFinish } from '@/app/api/jhipster-mocks/account/service';
 import {
   apiMethod,
   badRequestResponse,
-} from '@/app/api/jhipster-mocks/helpers';
+} from '@/app/api/jhipster-mocks/_helpers/api';
+import { db } from '@/app/api/jhipster-mocks/_helpers/db';
+import { formatUserFromDb } from '@/app/api/jhipster-mocks/_helpers/user';
 
 export const POST = apiMethod({
   public: true,
@@ -18,13 +20,33 @@ export const POST = apiMethod({
       .safeParse(await req.json());
 
     if (!bodyParsed.success) {
+      return badRequestResponse({ details: bodyParsed.error });
+    }
+
+    // Clear all expired tokens
+    await db.verificationToken.deleteMany({
+      where: { expires: { lt: new Date() } },
+    });
+
+    const verificationToken = await db.verificationToken.findUnique({
+      where: { token: bodyParsed.data.key },
+    });
+
+    if (!verificationToken) {
       return badRequestResponse();
     }
 
-    const user = await resetPasswordFinish({
-      token: bodyParsed.data.key,
-      newPassword: bodyParsed.data.newPassword,
-    });
+    const passwordHash = await bcrypt.hash(bodyParsed.data.newPassword, 12);
+
+    const [updatedUser] = await db.$transaction([
+      db.user.update({
+        where: { id: verificationToken.userId },
+        data: { password: passwordHash },
+      }),
+      db.verificationToken.delete({ where: { token: bodyParsed.data.key } }),
+    ]);
+
+    const user = formatUserFromDb(updatedUser);
 
     if (!user) {
       return badRequestResponse({ title: 'Invalid token' });
