+++ src/app/api/jhipster-mocks/activate/route.ts
@@ -1,10 +1,10 @@
 import { NextResponse } from 'next/server';
 
-import { activateAccount } from '@/app/api/jhipster-mocks/account/service';
 import {
   apiMethod,
   badRequestResponse,
-} from '@/app/api/jhipster-mocks/helpers';
+} from '@/app/api/jhipster-mocks/_helpers/api';
+import { db } from '@/app/api/jhipster-mocks/_helpers/db';
 
 export const GET = apiMethod({
   public: true,
@@ -15,7 +15,26 @@ export const GET = apiMethod({
       return badRequestResponse();
     }
 
-    const user = await activateAccount({ token });
+    // Clear all expired tokens
+    await db.verificationToken.deleteMany({
+      where: { expires: { lt: new Date() } },
+    });
+
+    const verificationToken = await db.verificationToken.findUnique({
+      where: { token },
+    });
+
+    if (!verificationToken) {
+      return undefined;
+    }
+
+    const [user] = await db.$transaction([
+      db.user.update({
+        where: { id: verificationToken.userId },
+        data: { activated: true },
+      }),
+      db.verificationToken.delete({ where: { token } }),
+    ]);
 
     if (!user) {
       return badRequestResponse();
