+++ src/app/api/jhipster-mocks/admin/users/[login]/route.ts
@@ -2,38 +2,45 @@ import { NextResponse } from 'next/server';
 import { z } from 'zod';
 
 import {
-  getUserByLogin,
-  removeUserByLogin,
-} from '@/app/api/jhipster-mocks/admin/users/service';
-import {
   apiMethod,
   badRequestResponse,
   notSignedInResponse,
-} from '@/app/api/jhipster-mocks/helpers';
+} from '@/app/api/jhipster-mocks/_helpers/api';
+import { db } from '@/app/api/jhipster-mocks/_helpers/db';
+import { formatUserFromDb } from '@/app/api/jhipster-mocks/_helpers/user';
 
 export const GET = apiMethod({
   admin: true,
   handler: async ({ params }) => {
     const safeParams = z.object({ login: z.string() }).safeParse(params);
     if (!safeParams.success) {
-      return badRequestResponse();
+      return badRequestResponse({ details: safeParams.error });
     }
-    const user = await getUserByLogin(safeParams.data.login);
+
+    const user = formatUserFromDb(
+      await db.user.findUnique({ where: { login: safeParams.data.login } })
+    );
+
     return NextResponse.json(user);
   },
 });
+
 export const DELETE = apiMethod({
   admin: true,
   handler: async ({ params, user }) => {
     if (!user?.id) {
       return notSignedInResponse();
     }
+
     const safeParams = z.object({ login: z.string() }).safeParse(params);
     if (!safeParams.success) {
-      return badRequestResponse();
+      return badRequestResponse({ details: safeParams.error });
     }
 
-    await removeUserByLogin(safeParams.data.login, user.id);
+    await db.user.delete({
+      where: { login: safeParams.data.login, NOT: { id: user.id } },
+    });
+
     return new NextResponse('ok', { status: 200 });
   },
 });
