+++ src/app/api/jhipster-mocks/authenticate/route.ts
@@ -1,11 +1,13 @@
+import bcrypt from 'bcrypt';
+import jwt from 'jsonwebtoken';
 import { NextResponse } from 'next/server';
 import { z } from 'zod';
 
-import { login } from '@/app/api/jhipster-mocks/authenticate/service';
 import {
   apiMethod,
   badRequestResponse,
-} from '@/app/api/jhipster-mocks/helpers';
+} from '@/app/api/jhipster-mocks/_helpers/api';
+import { db } from '@/app/api/jhipster-mocks/_helpers/db';
 
 export const POST = apiMethod({
   demo: 'allowed',
@@ -16,13 +18,27 @@ export const POST = apiMethod({
       .safeParse(await req.json());
 
     if (!bodyParsed.success) {
-      return badRequestResponse();
+      return badRequestResponse({ details: bodyParsed.error });
     }
 
-    const token = await login({
-      login: bodyParsed.data.username,
-      password: bodyParsed.data.password,
+    const user = await db.user.findUnique({
+      where: { login: bodyParsed.data.username },
     });
+
+    if (!user?.password || !user?.activated) {
+      return undefined;
+    }
+
+    const isPasswordValid = await bcrypt.compare(
+      bodyParsed.data.password,
+      user.password
+    );
+
+    if (!isPasswordValid) {
+      return undefined;
+    }
+
+    const token = await jwt.sign({ id: user.id }, process.env.AUTH_SECRET);
 
     if (!token) {
       return badRequestResponse();
