+++ src/app/api/jhipster-mocks/register/route.ts
@@ -1,12 +1,16 @@
+import bcrypt from 'bcrypt';
+import { randomUUID } from 'crypto';
+import dayjs from 'dayjs';
 import { NextResponse } from 'next/server';
 import { z } from 'zod';
 
-import { createAccount } from '@/app/api/jhipster-mocks/account/service';
 import {
   apiMethod,
   badRequestResponse,
   unknownErrorResponse,
-} from '@/app/api/jhipster-mocks/helpers';
+} from '@/app/api/jhipster-mocks/_helpers/api';
+import { db } from '@/app/api/jhipster-mocks/_helpers/db';
+import { prepareUserForDb } from '@/app/api/jhipster-mocks/_helpers/user';
 
 export const POST = apiMethod({
   public: true,
@@ -21,10 +25,36 @@ export const POST = apiMethod({
       .safeParse(await req.json());
 
     if (!bodyParsed.success) {
-      return badRequestResponse();
+      return badRequestResponse({ details: bodyParsed.error });
     }
 
-    const user = await createAccount(bodyParsed.data);
+    const passwordHash = await bcrypt.hash(bodyParsed.data.password, 12);
+
+    const user = await db.user.create({
+      data: prepareUserForDb({
+        email: bodyParsed.data.email.toLowerCase().trim(),
+        login: bodyParsed.data.login.toLowerCase().trim(),
+        password: passwordHash,
+        langKey: bodyParsed.data.langKey,
+      }),
+    });
+
+    const token = randomUUID();
+
+    await db.verificationToken.create({
+      data: {
+        userId: user.id,
+        token,
+        expires: dayjs().add(1, 'hour').toDate(),
+      },
+    });
+
+    // REPLACE ME WITH EMAIL SERVICE
+    console.log(`👇👇👇👇👇👇👇👇👇👇
+✉️ Activation link: ${process.env.NEXT_PUBLIC_BASE_URL}/app/account/activate?key=${token}
+👆👆👆👆👆👆👆👆👆👆`);
+
+    return user;
 
     if (!user) {
       return unknownErrorResponse();
